requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
python-decouple==3.6
python-dotenv==0.19.2
python-environ==0.4.54
sqlparse==0.4.2
tzdata==2021.5
